requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
grpcio-reflection==1.59.0
protobuf==4.24.4
numpy==1.26.0
scikit-learn==1.3.1
